# lib/api-zod/src/generated/types/appConfigResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AppConfigResponseConfigVars } from "./appConfigResponseConfigVars";

export interface AppConfigResponse {
  success: boolean;
  appName: string;
  configVars: AppConfigResponseConfigVars;
}