# lib/api-zod/src/generated/types/deployRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DeployRequestBotType } from "./deployRequestBotType";
import type { DeployRequestBotVars } from "./deployRequestBotVars";

export interface DeployRequest {
  botType: DeployRequestBotType;
  /**
   * @minLength 3
   * @maxLength 30
   */
  appName: string;
  botVars: DeployRequestBotVars;
}